Refactor Components Structure and Standardize File Names
