+++ src/pages/Boards/BoardContent/ListColumns/Column/Column.jsx
@@ -1,0 +1,149 @@
+import { useState } from 'react'
+import Typography from '@mui/material/Typography'
+import Box from '@mui/material/Box'
+import Menu from '@mui/material/Menu'
+import MenuItem from '@mui/material/MenuItem'
+import Button from '@mui/material/Button'
+import Divider from '@mui/material/Divider'
+import ListItemText from '@mui/material/ListItemText'
+import ContentCut from '@mui/icons-material/ContentCut'
+import Cloud from '@mui/icons-material/Cloud'
+import Tooltip from '@mui/material/Tooltip'
+
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
+import ListItemIcon from '@mui/material/ListItemIcon'
+import DeleteForeverIcon from '@mui/icons-material/DeleteForever'
+import ContentCopyIcon from '@mui/icons-material/ContentCopy'
+import ContentPasteIcon from '@mui/icons-material/ContentPaste'
+import AddCardIcon from '@mui/icons-material/AddCard'
+import DragHandleIcon from '@mui/icons-material/DragHandle'
+
+import ListCards from './ListCards/ListCards'
+
+function Column() {
+  const [anchorEl, setAnchorEl] = useState(null)
+  const open = Boolean(anchorEl)
+
+  const handleClick = (event) => {
+    setAnchorEl(event.currentTarget)
+  }
+
+  const handleClose = () => {
+    setAnchorEl(null)
+  }
+  return (
+    <Box
+      sx={{
+        maxWidth: '300px',
+        minWidth: '300px',
+        ml: 2,
+        borderRadius: '6px',
+        bgcolor: (theme) => (theme.palette.mode === 'dark' ? '#333643' : '#ebecf0'),
+        height: 'fit-content',
+        maxHeight: (theme) => `calc(${theme.trello.boardContentHeight} - ${theme.spacing(5)} )`
+      }}
+    >
+      {/* Box column HEADER */}
+      <Box
+        sx={{
+          height: (theme) => theme.trello.columnHeaderHeight,
+          p: 2,
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'space-between'
+        }}
+      >
+        <Typography
+          variant="h6"
+          sx={{
+            fontSize: '1rem',
+            cursor: 'pointer',
+            fontWeight: 'fontSupperBold'
+          }}
+        >
+          Column Title
+        </Typography>
+        <Box>
+          <Tooltip title="More Option">
+            <ExpandMoreIcon
+              sx={{ color: 'text.primary', cursor: 'pointer' }}
+              id="basic-column-dropdown"
+              aria-controls={open ? 'basic-menu-column-dropdown' : undefined}
+              aria-haspopup="true"
+              aria-expanded={open ? 'true' : undefined}
+              onClick={handleClick}
+            />
+          </Tooltip>
+          <Menu
+            id="basic-menu-column-dropdown"
+            anchorEl={anchorEl}
+            open={open}
+            onClose={handleClose}
+            MenuListProps={{ 'aria-labelledby': 'basic-column-dropdown' }}
+          >
+            <MenuItem>
+              <ListItemIcon>
+                <AddCardIcon fontSize="small" />
+              </ListItemIcon>
+              <ListItemText>Add new card</ListItemText>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <ContentCut fontSize="small" />
+              </ListItemIcon>
+              <ListItemText>Cut</ListItemText>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <ContentCopyIcon fontSize="small" />
+              </ListItemIcon>
+              <ListItemText>Copy</ListItemText>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <ContentPasteIcon fontSize="small" />
+              </ListItemIcon>
+              <ListItemText>paste</ListItemText>
+            </MenuItem>
+
+            <Divider />
+
+            <MenuItem>
+              <ListItemIcon>
+                <Cloud fontSize="small" />
+              </ListItemIcon>
+              <ListItemText>Archive this column</ListItemText>
+            </MenuItem>
+            <MenuItem>
+              <ListItemIcon>
+                <DeleteForeverIcon fontSize="small" />
+              </ListItemIcon>
+              <ListItemText>Remove this column</ListItemText>
+            </MenuItem>
+          </Menu>
+        </Box>
+      </Box>
+
+      {/* Box List Cards */}
+      <ListCards />
+
+      {/* Box column FOOTER */}
+      <Box
+        sx={{
+          height: (theme) => theme.trello.columnFooterHeight,
+          p: 2,
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'space-between'
+        }}
+      >
+        <Button startIcon={<AddCardIcon />}>Add new card</Button>
+        <Tooltip title="Drag to move">
+          <DragHandleIcon sx={{ cursor: 'pointer' }} />
+        </Tooltip>
+      </Box>
+    </Box>
+  )
+}
+
+export default Column
