+++ src/pages/Boards/_id.jsx
@@ -1,8 +1,7 @@
-// Boards detail
 import Container from '@mui/material/Container'
-import AppBar from '~/components/AppBar'
-import BoardBar from './BoardBar'
-import BoardContent from './BoardContent'
+import AppBar from '~/components/AppBar/AppBar'
+import BoardBar from './BoardBar/BoardBar'
+import BoardContent from './BoardContent/BoardContent'
 
 function Board() {
   return (
